+++ app/Http/Controllers/Bungadavi/BasicSetting/TimeSlotController.php
@@ -15,6 +15,7 @@ class TimeSlotController extends Controller
 
     public function index(TimeSlotDataTable $datatables)
     {
+        $this->authorize("view time slot");
         $data = [
             'title'         => 'Time Slot',
             'subtitle'      => 'Time Slot',
@@ -34,6 +35,7 @@ class TimeSlotController extends Controller
      */
     public function create()
     {
+        $this->authorize("create time slot");
         foreach(City::all() as $city){
             $city_selected[$city->id] = $city->name;
         }
@@ -85,6 +87,7 @@ class TimeSlotController extends Controller
 
     public function edit($id)
     {
+        $this->authorize("edit time slot");
         foreach(City::all() as $city){
             $city_selected[$city->id] = $city->name;
         }
@@ -128,6 +131,7 @@ class TimeSlotController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize("delete time slot");
         return TimeSlot::find($id)->delete();
     }
