+++ app/Http/Controllers/Bungadavi/User/GroupController.php
@@ -19,6 +19,7 @@ class GroupController extends Controller
      */
     public function index(GroupDataTable $datatables)
     {
+        $this->authorize("view user group");
         $data = [
             'title'         => 'User Group Position Management',
             'subtitle'      => 'User Group Position List',
@@ -37,6 +38,7 @@ class GroupController extends Controller
      */
     public function create()
     {
+        $this->authorize("create user group");
         $data = [
             'title'         => 'Form User Admin Management',
             'subtitle'      => 'User Admin Form',
@@ -87,6 +89,7 @@ class GroupController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize("edit user group");
         $data = [
             'title'         => 'Form User Admin Management',
             'subtitle'      => 'User Admin Form',
@@ -131,6 +134,7 @@ class GroupController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize("delete user group");
         return $position = Position::findOrFail($id)->delete();
     }
